+++ core/src/models/pre_check_query.rs
@@ -8,7 +8,7 @@ pub struct PreCheckoutQuery {
     pub id: String,
     pub from: User,
     pub currency: String,
-    pub total_amount: u32,
+    pub total_amount: u64,
     pub invoice_payload: String,
     pub shipping_option_id: Option<String>,
     pub order_info: Option<OrderInfo>,
