+++ core/src/models/proximity_alert_trigger.rs
@@ -6,5 +6,5 @@ use crate::models::user::User;
 pub struct ProximityAlertTriggered {
     pub traveler: User,
     pub watcher: User,
-    pub distance: u32,
+    pub distance: u64,
 }
