+++ core/src/models/video_chat.rs
@@ -4,7 +4,7 @@ use crate::models::user::User;
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
 pub struct VideoChatEnded {
-    pub duration: u32,
+    pub duration: u64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
