+++ core/src/models/video_note.rs
@@ -6,8 +6,8 @@ use crate::models::photo_size::PhotoSize;
 pub struct VideoNote {
     pub file_id: String,
     pub file_unique_id: String,
-    pub length: u32,
-    pub duration: u32,
+    pub length: u64,
+    pub duration: u64,
     pub thumbnail: Option<PhotoSize>,
     pub file_size: Option<u64>,
 }
