+++ core/src/params/webhook_param.rs
@@ -11,7 +11,7 @@ pub struct SetWebhookParams {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub ip_address: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_connections: Option<u32>,
+    pub max_connections: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub allowed_updates: Option<Vec<AllowedUpdate>>,
     #[serde(skip_serializing_if = "Option::is_none")]
